qc/picard: relocated to 'qc.apps.picard'.
